refactor feedback for better module organizationa and alter table
