+++ app.js
@@ -6,6 +6,7 @@ const connectDB = require("./db/connectDb")
 const authRoutes = require("./routes/authRoutes")
 const userRoutes = require("./routes/userRoutes")
 const sessionRoutes = require("./routes/sessionRoutes")
+const feedbackRoutes = require("./routes/feedbackRoutes")
 const cookiesParser = require("cookie-parser")
 
 const port = process.env.PORT
@@ -18,6 +19,7 @@ app.use(cookiesParser())
 app.use("/", authRoutes)
 app.use("/user", userRoutes)
 app.use("/sessions", sessionRoutes)
+app.use("/feedback", feedbackRoutes)
 
 connectDB()
