+++ controllers/userController.js
@@ -1,6 +1,5 @@
-const { v4: uuidv4 } = require("uuid")
 const response = require("../response")
-const { updateRecord, getRecordById, getAllRecords, insertRecord, checkRecordExists } = require("../utils/sqlFunctions")
+const { updateRecord, getRecordById, checkRecordExists } = require("../utils/sqlFunctions")
 
 const updateUserProfile = async (req, res) => {
   const { userid } = req.user
@@ -25,29 +24,6 @@ const getUserProfile = async (req, res) => {
   }
 }
 
-const viewAllSessions = async (req, res) => {
-  try {
-    const sessions = await getAllRecords("sessions")
-    response(200, sessions, "Sessions fetched successfully", res)
-  } catch (error) {
-    response(500, "", error, res)
-  }
-}
-
-const leaveFeedback = async (req, res) => {
-  const feedbackData = {
-    feedbackid: uuidv4(),
-    userid: req.user.userid,
-    ...req.body
-  }
-
-  try {
-    const newFeedback = await insertRecord("feedback", feedbackData)
-    response(201, newFeedback, "Feedback submitted successfully", res)
-  } catch (error) {
-    response(500, "", error, res)
-  }
-}
 
 const viewUserInfoByUsername = async (req, res) => {
   const { username } = req.params
@@ -64,10 +40,9 @@ const viewUserInfoByUsername = async (req, res) => {
   }
 }
 
+
 module.exports = {
   updateUserProfile,
   getUserProfile,
-  viewAllSessions,
-  leaveFeedback,
   viewUserInfoByUsername,
 }
