+++ schemas/sessionSchema.js
@@ -7,7 +7,9 @@ const sessionSchema =
     start_time TIMESTAMP,
     end_time TIMESTAMP,
     max_seats INT NOT NULL,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+    created_by VARCHAR ,
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    FOREIGN KEY (created_by) REFERENCES users(userid)
   )`
 
 module.exports = sessionSchema
